backend/requirements.txt
fastapi
uvicorn[standard]
nova-act
strands-agents[bidi]
boto3
aws-sdk-bedrock-runtime
pydantic-settings
